Restructure the project into subprojects
