+++ vaadin-notification-flow-demo/src/main/java/com/vaadin/flow/component/notification/demo/NotificationView.java
