+++ vaadin-notification-flow/src/main/java/com/vaadin/flow/component/notification/Notification.java
