+++ vaadin-notification-flow/src/test/java/com/vaadin/flow/component/notification/NotificationTest.java
@@ -31,6 +31,7 @@ import com.vaadin.flow.component.Component;
 import com.vaadin.flow.component.UI;
 import com.vaadin.flow.component.html.Div;
 import com.vaadin.flow.component.html.Label;
+import com.vaadin.flow.component.html.NativeButton;
 import com.vaadin.flow.component.notification.Notification.Position;
 import com.vaadin.flow.dom.Element;
 
@@ -43,6 +44,7 @@ import net.jcip.annotations.NotThreadSafe;
 public class NotificationTest {
 
     private UI ui = new UI();
+    private Notification notification;
 
     @Before
     public void setUp() {
@@ -209,6 +211,48 @@ public class NotificationTest {
     @Test(expected = IllegalArgumentException.class)
     public void addComponentAtIndex_indexIsBiggerThanChildrenCount() {
         addDivAtIndex(1);
+    }
+
+    @Test
+    public void stringAndDurationCtor() {
+        notification = new Notification("foo", 4000);
+        Assert.assertEquals(4000, notification.getDuration(), 0);
+        Assert.assertEquals("bottom-start",
+                notification.getPosition().getClientName());
+    }
+
+    @Test
+    public void stringDurAndPositionCtor() {
+        notification = new Notification("fooo", 10000, Position.TOP_END);
+        Assert.assertEquals(10000, notification.getDuration(), 0);
+        Assert.assertEquals("top-end",
+                notification.getPosition().getClientName());
+    }
+
+    @Test
+    public void componentCtor() {
+        notification = new Notification(new Label(), new NativeButton());
+
+        notification.setPosition(Position.BOTTOM_END);
+        Assert.assertEquals("bottom-end",
+                notification.getPosition().getClientName());
+    }
+
+    @Test
+    public void staticCtor() {
+        notification = Notification.show("fooooo", 4000,
+                Position.BOTTOM_CENTER);
+        Assert.assertEquals("bottom-center",
+                notification.getPosition().getClientName());
+    }
+
+    @Test
+    public void setPositon() {
+        notification = new Notification();
+
+        notification.setPosition(Position.BOTTOM_STRETCH);
+        Assert.assertEquals("bottom-stretch",
+                notification.getPosition().getClientName());
     }
 
     private void addDivAtIndex(int index) {
